+++ ext/nokogiri/xml_processing_instruction.c
@@ -22,7 +22,7 @@ new (int argc, VALUE *argv, VALUE klass)
 
   rb_scan_args(argc, argv, "3*", &document, &name, &content, &rest);
 
-  TypedData_Get_Struct(document, xmlDoc, &noko_xml_document_data_type, xml_doc);
+  xml_doc = noko_xml_document_unwrap(document);
 
   node = xmlNewDocPI(
            xml_doc,